S-13c extract ideabox to web/modules/ideas.js
Move all Ideas (Ötletláda) page code from app.js into a dedicated ES module
web/modules/ideas.js. The module exports loadIdeasPage and initIdeas; initIdeas
receives openModal/closeModal via dependency injection and registers window
globals for inline onclick handlers in template strings. Add
ideas-ui-contract.test.ts to guard the wiring.
